:truck: formatMMSS 함수 utils 폴더로 변경
