+++ src/hooks/mission/stopwatch/useStopwatch.ts
@@ -1,4 +1,5 @@
 import { useEffect, useState } from 'react';
+import { formatMMSS } from '@/utils/time';
 
 import { type StepType } from './useStopwatchStatus';
 
@@ -33,17 +34,3 @@ export default function useStopwatch(status: StepType) {
 
   return { minutes: formattedMinutes, seconds: formattedSeconds, stepper };
 }
-
-const formatMMSS = (second: number) => {
-  const minutes = Math.floor(second / 60); // 분 계산
-  const seconds = second % 60; // 초 계산
-  const formattedMinutes = String(minutes).padStart(2, '0'); // 두 자리로 변환
-  const formattedSeconds = String(seconds).padStart(2, '0'); // 두 자리로 변환
-
-  return {
-    minutes,
-    seconds,
-    formattedMinutes,
-    formattedSeconds,
-  };
-};
